added a very rough set of common instruments
I went with the largest possible ranges, but practical
ranges might be more useful.

Still a lot of work to do assigning clefs etc. I think that
complicated playability evaluation, like tympany and harp
tuning, is beyond the scope of this one-semester thesis. If
Heqet catches on, it would be best to have experts write the
code for their instruments anyway.