[p| ..pitch.. |] QQ suddenly stopped working...
either I imported something I hadn't, or a library got updated.
There was suddenly a "p" already in scope with type
Language.Haskell.TH.Lib.PatQ that was conflicted with my
Pitch QQ. As a quick fix, I changed my p to pp, but that's not
an obvious name choice. "pitch" is already in use at the moment
(although it should really be something else, as that's the lens
to get the "value" of a note)..